Apportionning

Calculates weights to distribute bird mortality between colonies and groups of colonies for a given species. For a given colony, these weights are influenced by: (1) Distance between the colony and the windfarms, (2) The population size of the species of interest for that colony and (3), the proportion of marine surface surrounding the colony.

bdy_apportionning(
  max_foraging_range_km,
  colonies,
  sea_area,
  tbl_dist,
  incl_pop_size = T,
  incl_sea_area = T
)

Arguments

max_foraging_range_km

numeric, maximum foraging range in kilometers for the species of interest

colonies

data frame with each row being a colony where the species of interest is present. Must have at least the following columns:

'group': numeric, colonies with the same number belong to the same cluster

'avg': numeric, average population size in that colony over the years of interests

'colony_code': character, unique identifier for the colony.

sea_area

named numeric vector giving for each colony the proportion of marine surface surrounding it. Each value should be named with the unique identifier for the corresponding colony.

tbl_dist

matrix giving the distances between colonies (rows) and wind farms (columns), as outputted by bdy_get_distances()

incl_pop_size

boolean, whether population size should influence the weight calculation

incl_sea_area

boolean, whether the proportion of marine surface surrounding the colony should influence the weight calculation

Value

List of 4 matrices:

'AW_colo' (rows=colonies, columns=windfarms): absolute weights for each colony/windfarm combination

'RW_colo' (rows=colonies, columns=windfarms): relative (normalized) weights for each colony/windfarm combination (sum of weights for a given windfarm = 1)

'AW_group' (rows=groups of colonies, columns=windfarms): absolute weights for each group/windfarm combination

'RW_group' (rows=groups of colonies, columns=windfarms): relative (normalized) weights for each group/windfarm combination (sum of weights for a given windfarm = 1)