Model with no impact

Computes the null population model (without impact from wind farms) for the birds species of interest

bdy_model_no_impact(
  count_data,
  PI,
  survival,
  fecundity,
  propRepro,
  modelFile,
  nimble = T,
  lightResults = T,
  ny_proj = 30,
  na = 5000,
  nb = 1000,
  ni = 20000,
  nc = 3,
  nt = 5
)

Arguments

count_data

matrix giving annual bird counts for each group of colonies (rows) and each year of interest (columns)

PI

matrix giving the proportion of colonies monitored for each group of colonies (rows) and each year of interest (columns)

survival

numeric vector giving the survival rates for the different age classes of the species of interest

fecundity

numeric vector giving the fecundity rates for the different age classes of the species of interest

propRepro

numeric vector giving the proportion of reproductive individual for the different age classes of the species of interest

modelFile

character, relative path for the text file containing the population model as jags code

nimble

boolean, whether to use nimble of jags to compute the bayesian model

lightResults

boolean, whether the output should only be the estimated output useful for the model with impact, or the full posterior distribution (can be quite heavy)

ny_proj

number of years from the last annual count for which to compute population projections

na

Number of iterations to run in the JAGS adaptive phase

nb

Number of iterations at the beginning of the chain to discard (i.e., the burn-in). Does not include the adaptive phase iterations.

ni

Total number of iterations per chain (including burn-in).

nc

Number of Markov chains to run.

nt

Thinning rate. Must be a positive integer.

Value

Data frame with nrow = ni.
If lightResults=T, columns contain estimated group size for the first year of projection (one column per group of colonies) and estimated annual growth rate (one column per group of colonies).
If lightRestuls=F, columns contain the full posterior distribution: different estimated output for each group of colonies, and each year of interest (count years + projection years).
Refer to the documentation of packages 'nimble' or 'jagsUI' for details.