Model with impact

Computes the population model with impact from wind farms for the birds species of interest

bdy_model_with_impact(
  n_group,
  ny_data,
  posterior,
  mortality,
  ni = 1000,
  ny_proj = 30,
  progress = NULL
)

Arguments

n_group

numeric, number of groups of colonies for the bird species of interest

ny_data

numeric, number of years for which count data exist for the bird species of interest

posterior

dataframe, estimated output from bdy_model_no_impact()

mortality

distribution of mortality accross groups of colonies, as outputted by bdy_process_mortality()

ni

number of iterations

ny_proj

number of years from the last annual count for which to compute population projections

progress

R shiny Progress object used to show the calculation progress to the user. Only for use within a shiny application.

Value

List of two arrays of size (n_group)(ny_data)(ni) containing projected population size:

sc0: projections with no impact from wind farms (null model).

sc1: projections with impacts from wind farms.