Process mortality file

Distributes national mortality among all colonies of a given species, based on weights calculated with bdy_apportionning()

bdy_process_mortality(collision, season, n_iteration = 1000, RW_group)

Arguments

collision

data frame giving mortality estimates for the species of interest and the wind farms of interest.Should have at least the following columns:

'windfarm': character, names of the windfarm for which the mortality due to collisions is estimated ('parc' is also accepted as a column name)

'month': numeric, month of the year (1 to 12) for which the mortality is estimated

'iteration': numeric, iteration index of the collision model.

'coefficient': numeric, estimated mortality coefficient from the collision model, for a given combination of windfarm, month and iteration

season

character vector of length 12, giving the presence status of the bird of interest on the french coasts for each month of the year, with:
'B' = breeding, 'R' = resident, 'T' = transition, 'M' = mixed, 'V' = visiting, 'A' = absent.
For more details refer to the Birdynamic report by Chambert et al.

RW_group

matrix (rows=groups of colonies, columns=windfarms) giving relative weights for each group/windfarm combination, with sum of weights for a given windfarm = 1, as outputed by bdy_apportionning()

Value

Matrix (rows = iterations, columns = groups of colonies) giving the distribution of mortality accross groups of colonies