Raw result tables

Compute result tables (national and by group of colonies)

bdy_raw_res_tables(mod_out)

Arguments

mod_out

Output of the bdy_model_with_impact() function, completed with colonies, mortality and distance data

Value

Three tables:

Simulated_National: Simulated counts for the 30 projected years at national level (used to plot temporal trends with bdy_plot_trends())

Tableau_Subpop: Summary table of results by colony (with relative impact, increase in probability of extinction, mortality, etc)

Tableau_National: Summary table of results at national level (with relative impact, increase in probability of extinction, mortality, etc)

Relative impact was calculated for each iteration as 100*(C0-C1)/C0where C0 is the estimated abundance of the species after 30 years without windfarm impact and C1 is the estimated abundance of the species after 30 years with windfarm impact (the median across simulations is provided, as well as the confidence interval). Increase in extinction probability is calculated as 100*(P1-P0) where P0 is the probability of extinction without windfarm impact (calculated as the proportion of simulations that reach counts of 0 after 30 years) and P1 is the probability of extinction with windfarm impact.