Run complete analysis

This function can be used to run the complete analysis (from calculating cost raster to getting model outputs). See the package vignette for more information.

bdy_run_analysis(
  species,
  countData,
  windfarms,
  timeRange = c(2009, 2021),
  foraging_ranges,
  mortality,
  n_iteration = 1000,
  ni_noImpact = 10000,
  ni_withImpact = 1000,
  ...
)

Arguments

species

vector of species names to include in the analysis

countData

table of bird counts

windfarms

sf object indicating the position of windfarms, for instance using bdydata_parcs_example; must be in EPSG 2154

timeRange

minimum and maximum years of bird counts to integrate in trends estimates

foraging_ranges

table of species foraging ranges, for instance using bdydata_foraging_ranges

mortality

table of collision from bdy_check_mortality(), for instance using bdydata_mortality_example

n_iteration

Number of iterations for the collision model (parameter used in bdy_process_mortality())

ni_noImpact

Number of iterations to use in bdy_model_no_impact()

ni_withImpact

Number of iterations to use in bdy_model_with_impact()

Value

List of model outputs for each species, including:

'no_impact': Bird counts assuming no impact from windfarm; see bdy_model_no_impact() for more details

'with_impact': Bird counts assuming no impact from windfarm; see bdy_model_with_impact() for more details

'colonies': Table of colonies where the species is present

'mortality': Table of collision per group of colony (see more details in bdy_process_mortality())

'distance': Table of distances between colonies and windfarms

'distanceBIN': Binarised table of distances between colonies and windfarms binarised, reporting whether the windfarm is accessible to the individuals from the colony (i.e., the distance is smaller than the maximum foraging distance of the species)